split a signle uber-jar into the core and apps executables to address the issue on Windows
